requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
ipykernel==6.12.1
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.1
jupyter-core==4.9.2
kiwisolver==1.4.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.1
packaging==21.3
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.1.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scipy==1.8.0
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.21
wcwidth==0.2.5
scikit-image==0.19.3